selftests/powerpc: Allocate base registers
Some ptrace selftests are passing input operands using a constraint that
can allocate any register for the operand, and using these registers on
load/store operations.

If the register allocated by the compile happens to be zero (r0), it might
cause a invalid memory address access (Segfault), since a load store
consider the address 0x0 if the register zero is the base register,
example:

	ld r2, 0(r1) /* will load into r2 the content of r1 address */
	ld r2, 0(r0) /* will load into r2 the context of 0x0 address */

In order to avoid this possible problem, the inline assembly constraint
should be aware tha these registers will be used as a base register, thus,
r0 is discarded and not allocated for this purpose.

Other than that, this patch removes operands that are not used on the
tests.

Signed-off-by: Breno Leitao <leitao@debian.org>