powerpc/tm: Restore transactional SPRs
In the previous TM code, trecheckpoint was being executed in the middle of
an exception, thus, SPRs were being restored to default kernel SPRs (as DSCR)
value after trecheckpoint was done.

With this current patchset, trecheckpoint is executed just before getting
to userspace, at ret_from_except_lite, for example. Thus, we do not need to
set default kernel DSCR value anymore, as we are leaving kernel space. The
value that be set is the same value that SPR was set when the process got
in kernel space, i.e. transactional value.

Now, when a treclaim happens, the current SPRs will be put at thread->dscr,tar,ppr
and the checkpointed SPRs are put at thread->tm_<spr>.

When recheckpointing, the thread->tm_<spr> values will be checkpointed and
transactional values will be loaded into the live register.

Signed-off-by: Breno Leitao <leitao@debian.org>