powerpc/tm: Restore facility after recheckpoint
After the facility registers are recheckpointed, they are clobbered with
the values that were recheckpointed (and are now in the checkpoint area).

If these registers will have the facility enabled on returning to
userspace, they need to be restored, otherwise userspace will see invalid
values.

This patch simplify the restore_tm_state() to just restore what is enabled
on the way to userspace.

Signed-off-by: Breno Leitao <leitao@debian.org>