Selftests/powerpc: Add a signal fuzzer selftest
This is a new selftest that raises a SIGUSR1 signal and handles it in a set
of different ways, trying to create different scenario for testing
purposes.

The process of raising a signal and calling sigreturn is interlaved with
TM operations, as starting, suspending and terminating a transaction. The
tests depends on random numbers, and, based on them, it sets different TM
state, that must be properly handled by the kernel.

Other than that, the test fills out the user context passed to the
sigreturn system call with random data, in order to make sure that the
signal handler syscall can handle different states properly.

This selftest has some parameters to control what kind of tests the user
wants to run, as for example, if a transation should be started prior to
signal being raised, or, after the signal being raised and before the
sigreturn.
If no parameter is given, the default is enabling all options.

This test does not check if the user context are being read and set
properly by the kernel. Its purpose, at this time, is basically
guaranteeing that the kernel does not crash on invalid scenarios.

Signed-off-by: Breno Leitao <leitao@debian.org>