io_uring: Pass whole sqe to commands
Currently rely on having large SQEs, but future operations might want to
use normal SQE.

We currently only pass part of the SQE (the payload) to the commands.
But, for commands that use normal SQE size, we want to access the
initial SQE fields inside the command callback.

This also extends async size to afford copying the whole double SQE size
(128 bytes).

Signed-off-by: Breno Leitao <leitao@debian.org>