ip6mr: Create a wrapper that understands ioctl
Create a wrapper for ip6mr that knows how to handle ioctl. The goal is to
move ioctl callback functions from accessing userspace, such as
put_user()/from_user(), from the function itself.

This function knows how to copy the proper structures to kernel space
memory before ioctl being called. This function also copy the buffer
that came from the function ioctl back to user space. That way, the
ioctl function does not need to deal with userspace memory and became
more flexible.

This function is currently not being called.

Signed-off-by: Breno Leitao <leitao@debian.org>