net: Create a wrapper for sock ioctl
Creates a wrapper for ioctls, where it call sock->ioctls() and save the
result back to userspace.

For ipmr and ip6mr cases (RAW and RAW6 proto), they are special since
they read the input from userspace and return a structure, other than
int.  This function handles these exceptions, and for everything else,
calls the ioctl and copy the result back to the userspace memory.

Signed-off-by: Breno Leitao <leitao@debian.org>