net: sock ioctl do not touch userspace memory
Moves the sock ioctls to avoid touching userspace memory. This changes
the ioctl format in the following way:

        int                     (*ioctl)(struct sock *sk, int cmd,
-                                        unsigned long arg);
+                                        int *karg);

So, the argument is a pointer to kernelspace memory, which usually
uses to save the result of the operation, and in the ip{6}mr case, it is
used as input buffer also (as output). The ip{6}mr case is handled by
sock_skprot_ioctl_ip{6}mr() function.

The wrapper that prepare the buffer and put the buffer back to user is
sock_skprot_ioctl(), so, instead of calling sk->sk_prot->ioctl(), the
calle now calles sock_skprot_ioctl().

Signed-off-by: Breno Leitao <leitao@debian.org>