net: ioctl: Use kernel memory on ioctl callbacks
Most of the ioctls to net protocols do not receive any "arg" input from
userspace, and returns an "int" back to userspace inside the
proto->ioctl callback. This is not flexible, because it is hard to reuse
these functions without passing userspace buffers. For example, in
io_uring and BPF.

Move the struct protocol ioctls to avoid touching userspace memory,
i.e., all protocol's ioctl callbacks is adapted to operate on a
kernel memory other than on userspace (so, no more {put,get}_user() and
friends being called in the ioctl callback.

This changes the ioctl format in the following way:

    int                     (*ioctl)(struct sock *sk, int cmd,
-                                        unsigned long arg);
+                                        int *karg);

So, the argument is a pointer to kernelspace memory, which usually
uses to save the result of the operation. For the common case, the
memory (karg) passed to to the ioctl callbacks is at
sock_skproto_ioctl_out(). This value will be copied back to userspace if
the ioctl() callback returns successfully.

The default case (where no input parameter is given, and an "int" is
returned to userspace) is the default case, but, there are three major
exceptions:

* Protocol RAW:
   * cmd = SIOCGETVIFCNT:
     * input and output = struct sioc_vif_req
   * cmd = SIOCGETSGCNT
     * input and output = struct sioc_sg_req

* Protocol RAW6:
   * cmd = SIOCGETMIFCNT_IN6
     * input and output = struct sioc_mif_req6
   * cmd = SIOCGETSGCNT_IN6
     * input and output = struct sioc_sg_req6

* Protocol PHONET:
  * cmd == SIOCPNADDRESOURCE | SIOCPNDELRESOURCE
     * input int (4 bytes)

 For the exception cases, functions sock_skproto_ioctl_in{out}() will
copy the userspace input, and copy it back to kernelspace.

The wrapper that prepare the buffer and put the buffer back to user is
sock_skprot_ioctl(), so, instead of calling sk->sk_prot->ioctl(), the
callee now calls sock_skprot_ioctl().

Signed-off-by: Breno Leitao <leitao@debian.org>